Bugfixes, update repo

+++ src/repositories/ExcursionRepository.cpp
@@ -42,7 +42,7 @@ Vector<Excursion> ExcursionRepository::mapToExcursions(const Vector<String>& row
 	for (unsigned int i = 0; i < rowsSize; i++)
 	{
 		Vector<String> data = rows[i].split(AppConfig::fileDelimiter);
-		res.pushBack(Excursion{ data[0], Date{data[1]}, Date{data[2]}, String::toInt(data[3]), data[4], data[5].split(AppConfig::vectorValuesDelimiter) });
+		res.pushBack(Excursion{ data[0], Date{data[1]}, Date{data[2]}, (unsigned short)String::toInt(data[3]), data[4], data[5].split(AppConfig::vectorValuesDelimiter) });
 	}
 
 	return res;
